requirements.txt
Django==5.2.11

djangorestframework==3.16.1
django-filter==25.2
drf-yasg==1.21.14
djangorestframework-simplejwt==5.5.1

django-cors-headers==4.9.0

psycopg2-binary==2.9.11

redis==7.2.0
django-redis==6.0.0

celery==5.6.2
django-celery-beat==2.8.1

sentry-sdk==2.56.0
lxml_html_clean
django-prometheus==2.4.1
prometheus-client==0.24.1
bcrypt
python-dotenv==1.2.1
flower==2.0.1

gunicorn

# AI / APIs
openai==2.28.0
groq==1.1.1

# News scraping
beautifulsoup4==4.14.3
lxml==6.0.2
newspaper3k==0.2.8
feedparser==6.0.12
tldextract==5.3.1

django-encrypted-model-fields
# Google APIs (YouTube)
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-oauthlib==1.3.0